==> ANALYSIS/4_IF-SitePred_predict_residues.ipynb <==
# Running IF-SitePred in Jupyter NB

## Imports

import os
import json
import esm
import torch
import pickle
import numpy as np 

## Utils functions

def save_to_pickle(variable, file_path):
    with open(file_path, 'wb') as file:
        pickle.dump(variable, file)

def read_from_pickle(file_path):
    with open(file_path, 'rb') as file:
        return pickle.load(file)

## Predicting residues functions

def get_chains(pdb_path):
    chains = [i[21:22] for i in open(pdb_path, 'r').readlines() if i.startswith('ATOM')]
    return list(set(chains))

def get_index_coords(pdb_path, chain):
    coords_resi = {}
    atoms = [i for i in open(pdb_path, 'r').readlines() if i.startswith('ATOM') and i[21:22] == chain]
    for atom in atoms:
        coords = [atom[30:38].strip(), atom[38:46].strip(), atom[46:55].strip()]
        coords_resi[','.join(coords)] = int(atom[22:26].strip())
    return coords_resi


def load_models(num_models):
    models = []
    for i in range(num_models):
        models.append(pickle.load(open(f'/Users/2394007/Documents/FROM_CLUSTER/CLUSTER/summer_project/binding-sites/models/lgbm_if1_{i}.pkl', 'rb')))
    return models

def write_chimerax_attr_file(data, attr_name, file_name, model_id='1', chain_id='A'):
    with open(file_name, 'w') as file:
        # Write the header
        file.write(f"attribute: {attr_name}\n")
        file.write("match mode: 1-to-1\n")
        file.write("recipient: residues\n")
        file.write("\n")  # Blank line for readability
        
        # Write each residue's attribute
        for res_num, attr_value in data.items():
            file.write(f"\t#{model_id}/{chain_id}:{res_num}\t{attr_value}\n")

## Load ESM-IF model

model, alphabet = esm.pretrained.load_model_and_alphabet_local('/Users/2394007/Documents/FROM_CLUSTER/CLUSTER/summer_project/binding-sites/ESM_models/esm_if1_gvp4_t16_142M_UR50.pt')
model.eval()

models = load_models(40)
model_num = len(models)

## Target selection

target_dir = './../../clean_rep_chains'

targets = read_from_pickle("./results/PDB_rep_chains_files_V2.pkl")

#targets = [os.path.join(target_dir, t) for t in target_names]

targets[:5]

print(len(targets))

for t in targets:
    if os.path.isfile(os.path.join(target_dir, t)):
        continue
    else:
        print(f'{t} does not exist')

preds_dir = "./results/IFSP_preds"

preds = os.listdir(preds_dir)

preds

N_targets = len(targets)
print(N_targets)

It is now 19:33 on 04/04/2024

It is now 22:19 on 08/04/2024

## Running IF-SitePred on Human AF

errors = []
done = 0
for i, target in enumerate(targets):

    if i % 50 == 0:
        print(i)

    target_id = target.split(".")[0]
    target_path = os.path.join(target_dir, target)

    prediction_dir = os.path.join(preds_dir, target_id)

    if not os.path.isdir(prediction_dir):
        os.mkdir(prediction_dir)
        
    lig_scores_out = os.path.join(prediction_dir, "{}_lig_scores.pkl".format(target_id))
    lig_label_out = os.path.join(prediction_dir, "{}_lig_labels.pkl".format(target_id))
    binding_ress_out = os.path.join(prediction_dir, "{}_binding_ress.pkl".format(target_id))
    ligandability_attr_out = os.path.join(prediction_dir, "{}_ligandability.defattr".format(target_id))
    lig_label_attr_out = os.path.join(prediction_dir, "{}_ligand_binding.defattr".format(target_id))

    if os.path.isfile(lig_scores_out) and os.path.isfile(lig_label_out) and os.path.isfile(binding_ress_out) and os.path.isfile(ligandability_attr_out) and os.path.isfile(lig_label_attr_out):
        #print("Results exist for {}".format(target_id))
        done += 1
        continue
    
    chains = sorted(get_chains(target_path))
    
    for chain in chains[:1]: # looping throuch chains (only using first)
    
        structure = esm.inverse_folding.util.load_structure(target_path, chain) # loading structure chain
        coords_resi = get_index_coords(target_path, chain) # dictionary where key is coordinates on a string and value is resNum
        coords, seq = esm.inverse_folding.util.extract_coords_from_structure(structure)	# coordinates and sequence in string
        if1 = esm.inverse_folding.util.get_encoder_output(model, alphabet, coords).detach().numpy() # generate embeddings from structure coordinates
    
        raw_preds = []
        round_preds = [] 
        for model_flaml in models:
            raw_p = model_flaml.predict(if1) # raw probs to 0/1
            raw_preds.append(raw_p)
            
            round_p = np.rint(raw_p)
            round_preds.append(round_p) # rounded to 0/1
    
        pred_intersection = []
        for i in range(len(round_preds[0])):
            pred_intersection.append(1 if [round_preds[a][i] for a in range(model_num)].count(1) == model_num else 0)
        
        res_nums = []
        #print(coords)
        for i in range(len(coords)):
            coord_str = ','.join([str("{:.3f}".format(x)) for x in coords[i][0]])
            if pred_intersection[i] == 1:
                try:
                    res_nums.append(str(coords_resi[coord_str]))
                except Exception as e:
                    print("ERROR with {}".format(target_id)) # it could be that alt locs make this crash
                    errors.append(target)
                    continue
    
        lig_scores = np.mean(raw_preds, axis = 0)
        res_score_dict = {} # key is resNum, value is ligandability score
        bin_lab_dict = {} # key is resNum, value is binary 0/1 label
        for i in range(len(coords)):
            coord_str = ','.join([str("{:.3f}".format(x)) for x in coords[i][0]])
            try:
                resnum = str(coords_resi[coord_str])
                res_score_dict[resnum] = round(lig_scores[i], 2)
                bin_lab_dict[resnum] = pred_intersection[i]
            except Exception as e:
                print("ERROR with {}".format(target_id)) # it could be that alt locs make this crash
                errors.append(target)
                continue
                
    
        
        save_to_pickle(
            res_score_dict,
            lig_scores_out
        )
        
        save_to_pickle(
            bin_lab_dict,
            lig_label_out
        )
        
        save_to_pickle(
            {chain: res_nums},
            binding_ress_out
        )
        
        write_chimerax_attr_file(
            res_score_dict, "ligandability",
            ligandability_attr_out
        )

        write_chimerax_attr_file(
            bin_lab_dict, "ligand_binding",
            lig_label_attr_out
        )
        print("There are {} ligand binding residues in chain {} of {}".format(str(len(res_nums)), chain, target_id))

It finished at 22:35 on 04/04/2024 182'

errors_unique = list(set(errors))
print(len(errors_unique))

## ChimeraX colouring commands

To colour by ligandability score:
    
    color byattribute r:ligandability #!1 target scab palette 0,white:0.5,#febe55:1,#de2d26

To colour by binary ligand-binding label:

    color byattribute r:ligand_binding #!1 target scab palette 0,white:0.5,white:1,red
