==> ANALYSIS/7.3_CONCAT_data_GrASP.ipynb <==
## Imports

from JSU_lib import *

## Functions

## Reading input data

grasp_pockets = read_from_pickle("./results/GrASP_pockets_dict_RIGHT_CLUSTERING.pkl")
grasp_pockets_scores = read_from_pickle("./results/GrASP_pocket_scores_dict_RIGHT_CLUSTERING.pkl")

len(grasp_pockets)

LIGYSIS_reps = read_from_pickle("./results/LIGYSIS_3448_chains.pkl")
len(set(LIGYSIS_reps).intersection(set(list(grasp_pockets.keys()))))

# Converting the nested dictionary to a format suitable for DataFrame
rows = []
for rep_chain, ids_sites in grasp_pockets.items():
    for ID, sites in ids_sites.items():
        #print(grasp_pockets_scores[rep_chain])
        try:
            rankings = {k: rank for rank, (k, v) in enumerate(sorted(grasp_pockets_scores[rep_chain].items(), key=lambda item: item[1], reverse=True), start=1)}
            rows.append({'rep_chain': rep_chain, 'ID': ID, 'site_ress': sites, 'score': grasp_pockets_scores[rep_chain][ID], 'RANK': rankings[ID]})
        except:
            print(grasp_pockets[rep_chain])
            print(grasp_pockets_scores[rep_chain])
            print(f'ERROR with {rep_chain} {ID}')
    #break

# Creating the DataFrame
grasp_pockets_df = pd.DataFrame(rows)
grasp_pockets_df["n_ress"] = grasp_pockets_df['site_ress'].apply(lambda x: len(x))
grasp_pockets_df = grasp_pockets_df.sort_values(by=["rep_chain", "RANK"]).reset_index(drop = True)

grasp_pockets_df.head(3)

for _, row in grasp_pockets_df.query('rep_chain == "5emn_B"').iterrows():
    ress_sels = ",".join(row.site_ress)
    print(f'{row.score}\tsel : {ress_sels}')

## Calcualte RoG of GrASP pockets

target_dir = './../DATA/trans_rep_chains'

rep_chains_sites = grasp_pockets_df.rep_chain.unique().tolist()

len(rep_chains_sites)

cc = [
    'label_seq_id_full', 'label_comp_id', 'type_symbol',
    'label_atom_id', 'Cartn_x', 'Cartn_y', 'Cartn_z'
]

T0 = time.time()
RoG_dict = {}
centres_dict = {}
for i, rep_chain in enumerate(rep_chains_sites):
    if i % 100 == 0:
        print(i)
    pdb_path = os.path.join(target_dir, f'{rep_chain}.trans.pdb') # calculating RoG in TRANS structures
    df = PDBXreader(inputfile=pdb_path).atoms(format_type="pdb", excluded=())[cc].query('label_atom_id == "CA"').reset_index(drop = True)
    #df.label_seq_id = df.label_seq_id.astype(int)
    rep_chain_df = grasp_pockets_df.query('rep_chain == @rep_chain')
    #print(len(acc_sites_df))
    for _, row in rep_chain_df.iterrows():
        site_ress = row.site_ress
        #print(site_ress)
        ress_CAs = df.query('label_seq_id_full in @site_ress')
        CAs_coords = np.array(list((zip(ress_CAs.Cartn_x, ress_CAs.Cartn_y, ress_CAs.Cartn_z))))
        center_of_mass = np.mean(CAs_coords, axis=0)
        distances_squared = np.sum((CAs_coords - center_of_mass)**2, axis=1)
        radius_of_gyration = round(np.sqrt(np.mean(distances_squared)), 2)
        RoG_dict[(rep_chain, row["RANK"])] = radius_of_gyration
        centres_dict[(rep_chain, row["RANK"])] = tuple([round(v, 3) for v in center_of_mass])
TF = time.time()

dTs = TF-T0

dTm = dTs/60

N_RoG = len(RoG_dict)

print(N_RoG) #8051

print(round(dTm,2))

print(round(dTs/N_RoG, 5))

round((N_RoG*(dTs/N_RoG))/60, 5) # 6 minutes

save_to_pickle(RoG_dict, "./results/GrASP_RoG_TRANS_RIGHT_CLUSTERING.pkl")

len(grasp_pockets_df) #used to be 8051

grasp_pockets_df["RoG"] = grasp_pockets_df.set_index(['rep_chain', 'RANK']).index.map(RoG_dict)

centres = []
errs = 0
for i, row in grasp_pockets_df.iterrows():
    try:
        centre = centres_dict[(row.rep_chain, row.RANK)]
        centres.append(centre)
    except:
        errs += 1
        centres.append(np.nan)
print(errs)

grasp_pockets_df["centre"] = centres

set(RoG_dict.keys()) == set(centres_dict.keys())

len(set(LIGYSIS_reps).intersection(set(list(grasp_pockets_df.rep_chain.unique()))))

grasp_pockets_df.head(3)

grasp_pockets_df.query('centre != centre')

#grasp_pockets_df.to_pickle("./results/GRASP_pockets_RoG_V2.pkl")

## Mapping PDB residues to UP residues

up_aas = {}

errors = []

for _, row in grasp_pockets_df.iterrows():
    b = 0
    rep_chain = row.rep_chain
    site_ress = row.site_ress
    rank = row["RANK"]
    mapping_dict = read_from_pickle(f'./../DATA/mappings_label_full/{rep_chain}.pkl')
    site_up_aas = []
    for res in site_ress:
        try:
            site_up_aas.append(mapping_dict[str(res)]) # I think there are integers here as it is not working with AltLocs, so none are present
        except:
            errors.append(rep_chain)
            b = 1
            break
    if b ==1:
        continue
    up_aas[(rep_chain, rank)] = site_up_aas

len(up_aas) #7948

len(errors) #103

errors = list(set(errors)) # 93 chains have issues with SIFTS mapping, e.g, 1nav_A. For some reason, the last 40 residues lack residue mapping to UniProt
print(len(errors)) #93

save_to_pickle(errors, "./results/GRaSP_no_sifts_TRANS_RIGHT_CLUSTERING.pkl")

grasp_pockets_df_w_sifts = grasp_pockets_df.query('rep_chain not in @errors').copy().reset_index(drop = True)
grasp_pockets_df_w_sifts["up_aas"] = grasp_pockets_df_w_sifts.set_index(['rep_chain', 'RANK']).index.map(up_aas)

grasp_pockets_df_w_sifts["n_up_aas"] = grasp_pockets_df_w_sifts['up_aas'].apply(lambda x: len(x))
assert grasp_pockets_df_w_sifts.n_ress.equals(grasp_pockets_df_w_sifts.n_up_aas)
grasp_pockets_df_w_sifts.drop(columns = ["n_up_aas"], inplace = True)

grasp_pockets_df_w_sifts.rename(columns = {"site_ress": "aas", "n_ress": "n_aas"}, inplace = True)

grasp_pockets_df_w_sifts.head(5)

len(grasp_pockets_df_w_sifts.rep_chain.unique()) #3937

len(set(LIGYSIS_reps).intersection(set(list(grasp_pockets_df_w_sifts.rep_chain.unique())))) #3353

grasp_pockets_df_w_sifts.to_pickle("./results/GRASP_pockets_sifted_RIGHT_CLUSTERING.pkl")

## Add pocket surfaces and volumes

#grasp_pockets_df_w_sifts = pd.read_pickle("./results/GRASP_pockets_sifted_V2.pkl")

master_SASA_dict = read_from_pickle("./results/master_SASA_dict.pkl")

grasp_pockets_df_w_sifts_SASA = calculate_total_sasa(grasp_pockets_df_w_sifts, master_SASA_dict)

grasp_pockets_df_w_sifts_SASA.query('SASA != SASA')

vols_dict = read_from_pickle("./results/GrASP_volumes_dict_RIGHT_CLUSTERING.pkl")

grasp_pockets_df_w_sifts_SASA["VOL"] = grasp_pockets_df_w_sifts_SASA.set_index(['rep_chain', 'ID']).index.map(vols_dict)

grasp_pockets_df_w_sifts_SASA

grasp_pockets_df_w_sifts_SASA.to_pickle("./results/GRASP_pockets_DEF_TRANS_RIGHT_CLUSTERING.pkl")

bad_reps = read_from_pickle("./results/OLD/other_errrors_reps.pkl")

grasp_pockets_df_w_sifts_SASA.query('VOL != VOL & rep_chain not in @bad_reps')

for _i, row in combined_SASA_df.iterrows():
    RoG = row.RoG
    rep_chain = row.rep_chain
    aas = row.aas
    ID = row.ID
    if 25<RoG<30:
        print(f"{rep_chain}_{ID}: {RoG} --> {'+'.join(aas)}")
