==> ANALYSIS/7.9_CONCAT_data_PocketFinder.ipynb <==
from JSU_lib import *

POCKETFINDER_dir = "./../DATA/pocketfinder_preds"

rep_chains_dir = "./../DATA/clean_rep_chains"

prot_pdb_ter = "_residue.pdb"
pocket_pdb_ter = "_pocket.pdb"
scores_ter = ".scores"
pymol_ter = ".pml"
dx_ter = ".dx"

POCKETFINDER_files = os.listdir(POCKETFINDER_dir)
pocket_files = sorted([el for el in POCKETFINDER_files if el.endswith(pocket_pdb_ter)])
score_files = sorted([el for el in POCKETFINDER_files if el.endswith(scores_ter)])
protein_files = sorted([el for el in POCKETFINDER_files if el.endswith(prot_pdb_ter)])

print(len(pocket_files))
print(len(score_files))
print(len(protein_files))

## Generating .ds file for PRANK re-scoring

header = """
# Dataset for rescoring ConCavity predictions

PARAM.PREDICTION_METHOD=concavity

HEADER: prediction protein

"""
# for i, file in enumerate(pocket_files):
#     rep_chain = file.split(".")[0]
#     print(f'./../../DATA/pocketfinder_preds/{file} ./../../DATA/clean_rep_chains/{rep_chain}.clean.pdb')

# running PRANK on it took 1 minutes 20.327 seconds with -threads 8

## Extracting pocket data

scores_master_dict = {}
rows = []
for i in range(len(pocket_files)):

    if i %100 == 0:
        print(i)
    pocket_file = pocket_files[i]
    score_file = score_files[i]
    rep_chain = pocket_file.split(".")[0]
    
    score_path = os.path.join(POCKETFINDER_dir, score_file)
    pocket_path = os.path.join(POCKETFINDER_dir, pocket_file)
    clean_pdb_path = f'./../DATA/clean_rep_chains/{rep_chain}.clean.pdb'
    
    clean_df = PDBXreader(inputfile=clean_pdb_path).atoms(format_type="pdb", excluded=())
    pocket_df = PDBXreader(inputfile=pocket_path).atoms(format_type="pdb", excluded=(), remove_hydrogens = False) # poitns stored as H HETATMS
    
    scores_dict = parse_concavity_residue_scores(score_path)
    scores_master_dict[rep_chain] = scores_dict

    coords_dict = create_concavity_coordinates_dict(pocket_df) # dict pocket ID: pocket grid points
    
    centroids_dict = calculate_concavity_centroids(coords_dict) # dict pocket ID: pocket centroids

    pocket_score_dict = calculate_concavity_pocket_score(pocket_df) # dict pocket ID: pocket score (SS of grid point scores)

    pocket_ress_dict = get_concavity_pocket_ress(coords_dict, clean_df)

    for k in centroids_dict.keys():
        #print(k)
        pocket_centroid = tuple([round(el, 4) for el in centroids_dict[k]])
        pocket_score = pocket_score_dict[k]
        pocket_ress = pocket_ress_dict[k]
        n_ress = len(pocket_ress)
        n_grid_points = len(coords_dict[k])
        
        pocket_ress_df = clean_df.query('label_seq_id_full in @pocket_ress')
        ress_CAs = pocket_ress_df.query('label_atom_id == "CA"')
        CAs_coords = np.array(list((zip(ress_CAs.Cartn_x, ress_CAs.Cartn_y, ress_CAs.Cartn_z))))
        center_of_mass = np.mean(CAs_coords, axis=0)
        distances_squared = np.sum((CAs_coords - center_of_mass)**2, axis=1)
        radius_of_gyration = round(np.sqrt(np.mean(distances_squared)), 2)
        d_row = [rep_chain, int(k)+1, n_grid_points, n_ress, radius_of_gyration, pocket_centroid, pocket_ress, pocket_score]
        rows.append(d_row)

    #break

POCKETFINDER_pockets = pd.DataFrame(rows, columns = ["rep_chain", "ID", "n_points", "n_aas", "RoG", "centre", "aas", "score"])

POCKETFINDER_pockets

print(len(POCKETFINDER_pockets))

## Mapping PDB residues to UP residues

up_aas = {}

errors = []

for _, row in POCKETFINDER_pockets.iterrows():
    b = 0
    rep_chain = row.rep_chain
    site_ress = row.aas
    ID = row["ID"]
    mapping_dict = read_from_pickle(f'./../DATA/mappings_label_full/{rep_chain}.pkl')
    site_up_aas = []
    for res in site_ress:
        try:
            site_up_aas.append(mapping_dict[str(res)]) # I think there are integers here as it is not working with AltLocs, so none are present
        except:
            errors.append(rep_chain)
            b = 1
            break
    if b ==1:
        continue
    up_aas[(rep_chain, ID)] = site_up_aas

print(len(up_aas)) #10227
print(len(errors)) #277

errors = list(set(errors)) # 242 chains have issues with SIFTS mapping
print(len(errors)) #242

save_to_pickle(errors, "./results/POCKETFINDER_no_sifts.pkl")

POCKETFINDER_pockets_w_sifts = POCKETFINDER_pockets.query('rep_chain not in @errors').copy().reset_index(drop = True)
POCKETFINDER_pockets_w_sifts["up_aas"] = POCKETFINDER_pockets_w_sifts.set_index(['rep_chain', 'ID']).index.map(up_aas)

POCKETFINDER_pockets_w_sifts["n_up_aas"] = POCKETFINDER_pockets_w_sifts['up_aas'].apply(lambda x: len(x))
assert POCKETFINDER_pockets_w_sifts.n_aas.equals(POCKETFINDER_pockets_w_sifts.n_up_aas)
POCKETFINDER_pockets_w_sifts.drop(columns = ["n_up_aas"], inplace = True)

POCKETFINDER_pockets_w_sifts

## Add pocket surfaces and volumes

master_SASA_dict = read_from_pickle("./results/master_SASA_dict.pkl")

POCKETFINDER_pockets_w_sifts_SASA = calculate_total_sasa(POCKETFINDER_pockets_w_sifts, master_SASA_dict)

vols_dict = read_from_pickle("./results/POCKETFINDER_volumes_dict.pkl")
POCKETFINDER_pockets_w_sifts_SASA["VOL"] = POCKETFINDER_pockets_w_sifts_SASA.set_index(['rep_chain', 'ID']).index.map(vols_dict)
print(len(POCKETFINDER_pockets_w_sifts_SASA.query('VOL != VOL')))
POCKETFINDER_pockets_w_sifts_SASA['centre_mat'] = POCKETFINDER_pockets_w_sifts_SASA.apply(apply_rotation, axis=1)

### Default RANK is unranked, RANK = ID, as LIGSITE did not assign scores nor ranked their pockets

The score here reported has been calculated by me, Javier Sánchez Utgés, in the same manner as done in P2Rank and similarly GrASP, as the sum of squares of grid point scores.

POCKETFINDER_pockets_w_sifts_SASA['RANK'] = POCKETFINDER_pockets_w_sifts_SASA.ID

POCKETFINDER_pockets_w_sifts_SASA

#POCKETFINDER_pockets_w_sifts_SASA['RANK'] = POCKETFINDER_pockets_w_sifts_SASA.groupby('rep_chain')['score'].rank(ascending=False, method='min').astype(int)

POCKETFINDER_pockets_w_sifts_SASA.to_pickle("./results/POCKETFINDER_pockets_DEF_TRANS_UNRANKED.pkl") # rank here is default (no ranking, i.e., RANK = ID)
