==> ANALYSIS/10_COMBINED_dataframe_generation.ipynb <==
## Imports

from JSU_lib import *

## Reading input data

LIGYSIS_sites = pd.read_pickle("./results/LIGYSIS_sites_DEF_TRANS.pkl")                                          # 2024
VNEGNN_pockets = pd.read_pickle("./results/VNEGNN_pockets_DEF_TRANS.pkl")                                        # 2024
IFSP_pockets = pd.read_pickle("./results/IFSP_pockets_DEF_TRANS.pkl")                                            # 2024
GRASP_pockets = pd.read_pickle("./results/GRASP_pockets_DEF_TRANS_RIGHT_CLUSTERING.pkl")                         # 2024

PURESNET_PRANK_CONS_pockets = pd.read_pickle("./results/puresnet_pranked_CONS_pockets_DEF_TRANS.pkl")            # 2021 + 2019
PURESNET_PRANK_pockets = pd.read_pickle("./results/puresnet_pranked_pockets_DEF_TRANS.pkl")                      # 2021 + 2015
PURESNET_pockets = pd.read_pickle("./results/PURESNET_pockets_DEF_TRANS.pkl")                                    # 2021

DEEPPOCKET_pockets = pd.read_pickle("./results/DeepPocket_pockets_DEF_TRANS2.pkl")                               # 2021

P2RANK_CONS_pockets = pd.read_pickle("./results/P2Rank_CONS_pockets_DEF_TRANS.pkl")                              # 2019
P2RANK_pockets = pd.read_pickle("./results/P2Rank_pockets_DEF_TRANS.pkl")                                        # 2018

FPOCKET_PRANK_CONS_pockets = pd.read_pickle("./results/fpocket_pranked_CONS_pockets_DEF_TRANS.pkl")              # 2009 + 2019
FPOCKET_PRANK_pockets = pd.read_pickle("./results/fpocket_pranked_pockets_DEF_TRANS.pkl")                        # 2009 + 2015
FPOCKET_pockets = pd.read_pickle("./results/fpocket_pockets_DEF_TRANS.pkl")                                      # 2009

POCKETFINDER_PRANK_CONS_pockets = pd.read_pickle("./results/pocketfinder_pranked_CONS_pockets_DEF_TRANS.pkl")    # 2005 + 2019
POCKETFINDER_PRANK_pockets = pd.read_pickle("./results/pocketfinder_pranked_pockets_DEF_TRANS.pkl")              # 2005 + 2015
POCKETFINDER_pockets = pd.read_pickle("./results/POCKETFINDER_pockets_DEF_TRANS_UNRANKED.pkl")                   # 2005

LIGSITE_PRANK_CONS_pockets = pd.read_pickle("./results/ligsite_pranked_CONS_pockets_DEF_TRANS.pkl")              # 1997 + 2019    
LIGSITE_PRANK_pockets = pd.read_pickle("./results/ligsite_pranked_pockets_DEF_TRANS.pkl")                        # 1997 + 2015    
LIGSITE_pockets = pd.read_pickle("./results/LIGSITE_pockets_DEF_TRANS_UNRANKED.pkl")                             # 1997

SURFNET_PRANK_CONS_pockets = pd.read_pickle("./results/surfnet_pranked_CONS_pockets_DEF_TRANS.pkl")              # 1995 + 2019
SURFNET_PRANK_pockets = pd.read_pickle("./results/surfnet_pranked_pockets_DEF_TRANS.pkl")                        # 1995 + 2015
SURFNET_pockets = pd.read_pickle("./results/SURFNET_pockets_DEF_TRANS_UNRANKED.pkl")                             # 1995


def get_sites_per_prot_LIGYSIS(df):
    a = df['rep_chain'].value_counts()
    sites_per_prot_df = pd.DataFrame(list(a.items()), columns=['key', 'sites_per_prot'])
    sites_per_prot_df[['origin', 'rep_chain']] = pd.DataFrame(sites_per_prot_df['key'].tolist(), index=sites_per_prot_df.index)
    sites_per_prot_df.drop('key', axis=1, inplace=True)
    sites_per_prot_df = sites_per_prot_df[['origin', 'rep_chain', 'sites_per_prot']]  # Reorder columns
    return df

len(LIGYSIS_sites['rep_chain'].value_counts().to_dict())

sites_per_prot = get_sites_per_prot_LIGYSIS(LIGYSIS_sites)

LIGYSIS_sites_per_prot

## Adding RANK = ID, as PUResNet does not score pockets

PURESNET_pockets['score'] = PURESNET_pockets['n_aas']

PURESNET_pockets['RANK'] = PURESNET_pockets.ID

## Re-formatting DeepPocket to consider all the re-scored fpocket candidates

DEEPPOCKET_rescored_pockets = DEEPPOCKET_pockets.copy()

DEEPPOCKET_rescored_pockets.drop(
    columns=[
        'score_fpocket', 'drug_score', 'n_alpha_spheres', 'total_sasa',
        'polar_sasa', 'apolar_sasa', 'volume_fpocket', 'mean_hydro_density',
        'mean_alpha_sphere_r', 'mean_alpha_sphere_solvent_acc',
        'apolar_alpha_sphere_prop', 'hydro_score', 'vol_score',
        'polarity_score', 'charge_score', 'prop_polar_atoms',
        'alpha_sphere_density', 'max_dist', 'flexibility', 'RANK_fpocket',
        'centre_deep1', 'centre_fpocket', 'aas_deep', 'n_aas_deep',
        'centre_deep2', 'centre_fpocket_trans', 'RoG_deep',
        'centre_deep_trans', 'centre_mat', 'distance',
        'barycentre_distance', 'centre_fpocket_mat', 'dist_fpocket_mat',
        'up_aas_deep', 'SASA', 'VOL'
    ],
    inplace=True
)

DEEPPOCKET_rescored_pockets.rename(
    columns = {
        "RANK_deep": "RANK",
        "score_deep": "score",
        "n_aas_fpocket": "n_aas",
        "aas_fpocket": "aas",
        "up_aas_fpocket": "up_aas",  
        "RoG_fpocket": "RoG",
        "barycentre_mat": "centre",
    },
    inplace = True
)

### Re-formatting DeepPocket dataframe to drop those fpocket pockets that were not segmented

DEEPPOCKET_segmented_pockets = DEEPPOCKET_pockets.query('n_aas_deep == n_aas_deep').copy().reset_index(drop = True)

DEEPPOCKET_segmented_pockets.drop(
    columns=[
        'score_fpocket', 'drug_score', 'n_alpha_spheres', 'total_sasa',
        'polar_sasa', 'apolar_sasa', 'volume_fpocket', 'mean_hydro_density',
        'mean_alpha_sphere_r', 'mean_alpha_sphere_solvent_acc',
        'apolar_alpha_sphere_prop', 'hydro_score', 'vol_score',
        'polarity_score', 'charge_score', 'prop_polar_atoms',
        'alpha_sphere_density', 'max_dist', 'flexibility',
        'RANK_fpocket','aas_fpocket', 'n_aas_fpocket', 'centre_fpocket', 'RoG_fpocket',  'up_aas_fpocket',
        'centre_deep1', 'centre_deep2','centre_deep_trans','centre_fpocket_trans', 'distance'
       ],
    inplace=True
)

DEEPPOCKET_segmented_pockets.rename(
    columns = {
        "RANK_deep": "RANK",
        "score_deep": "score",
        "aas_deep": "aas",
        "n_aas_deep": "n_aas",
        "centre_mat": "centre",
        "RoG_deep": "RoG",
        "up_aas_deep": "up_aas",        
    },
    inplace = True
)

DEEPPOCKET_segmented_pockets.n_aas = DEEPPOCKET_segmented_pockets.n_aas.astype(int)

DEEPPOCKET_segmented_pockets['RANK'] = DEEPPOCKET_segmented_pockets.groupby('rep_chain')['score'].rank(ascending=False, method='first').astype(int)

## Re-formatting other dataframes to take into account the correct centroid coordinates

## VN-EGNN ##
VNEGNN_pockets.drop(
    columns=[
        'centre', 'centre_trans',
       ],
    inplace=True
)

VNEGNN_pockets.rename(
    columns = {
        "centre_mat": "centre",
    },
    inplace = True
)

## IF-SITEPRED ##
IFSP_pockets.drop(
    columns=[
        'centre', 'centre_trans',
       ],
    inplace=True
)

IFSP_pockets.rename(
    columns = {
        "centre_mat": "centre",
    },
    inplace = True
)

## PURESNET ##
PURESNET_pockets.drop(
    columns=[
        'centre',
       ],
    inplace=True
)

PURESNET_pockets.rename(
    columns = {
        "centre_mat": "centre",
    },
    inplace = True
)

## P2RANK+CONS ##
P2RANK_CONS_pockets.drop(
    columns=[
       'centre_trans',
       ],
    inplace=True
)

## P2RANK ##
P2RANK_pockets.drop(
    columns=[
        'centre', 'centre_trans',
       ],
    inplace=True
)

P2RANK_pockets.rename(
    columns = {
        "centre_mat": "centre",
    },
    inplace = True
)

## POCKETFINDER ##
POCKETFINDER_pockets.drop(
    columns=[
        'centre'
       ],
    inplace=True
)

POCKETFINDER_pockets.rename(
    columns = {
        "centre_mat": "centre",
    },
    inplace = True
)

## LIGSITE ##
LIGSITE_pockets.drop(
    columns=[
        'centre'
       ],
    inplace=True
)

LIGSITE_pockets.rename(
    columns = {
        "centre_mat": "centre",
    },
    inplace = True
)

## SURFNET ##
SURFNET_pockets.drop(
    columns=[
        'centre'
       ],
    inplace=True
)

SURFNET_pockets.rename(
    columns = {
        "centre_mat": "centre",
    },
    inplace = True
)

## Formatting dataframe

LIGYSIS_sites["origin"] = "LIGYSIS"
VNEGNN_pockets["origin"] = "VN-EGNN"
IFSP_pockets["origin"] = "IF-SitePred"
GRASP_pockets["origin"] = "GrASP"

PURESNET_PRANK_CONS_pockets["origin"] = "PUResNet+PRANK+Cons"
PURESNET_PRANK_pockets["origin"] = "PUResNet+PRANK"
PURESNET_pockets["origin"] = "PUResNet"

DEEPPOCKET_segmented_pockets["origin"] = "DeepPocket-Segmented"
DEEPPOCKET_rescored_pockets["origin"] = "DeepPocket-Rescored"

P2RANK_CONS_pockets["origin"] = "P2Rank+Cons"
P2RANK_pockets["origin"] = "P2Rank"

FPOCKET_PRANK_pockets["origin"] = "fpocket+PRANK"
FPOCKET_PRANK_CONS_pockets["origin"] = "fpocket+PRANK+Cons"
FPOCKET_pockets["origin"] = "fpocket"

POCKETFINDER_PRANK_CONS_pockets["origin"] = "PocketFinder+PRANK+Cons"
POCKETFINDER_PRANK_pockets["origin"] = "PocketFinder+PRANK"
POCKETFINDER_pockets["origin"] = "PocketFinder"

LIGSITE_PRANK_CONS_pockets["origin"] = "Ligsite+PRANK+Cons"
LIGSITE_PRANK_pockets["origin"] = "Ligsite+PRANK"
LIGSITE_pockets["origin"] = "Ligsite"

SURFNET_PRANK_CONS_pockets["origin"] = "Surfnet+PRANK+Cons"
SURFNET_PRANK_pockets["origin"] = "Surfnet+PRANK"
SURFNET_pockets["origin"] = "Surfnet"

LIGYSIS_sites['up_aas'] = LIGYSIS_sites['up_aas'].apply(sorted)
VNEGNN_pockets['up_aas'] = VNEGNN_pockets['up_aas'].apply(sorted)
IFSP_pockets['up_aas'] = IFSP_pockets['up_aas'].apply(sorted)
GRASP_pockets['up_aas'] = GRASP_pockets['up_aas'].apply(sorted)
PURESNET_pockets['up_aas'] = PURESNET_pockets['up_aas'].apply(sorted)
DEEPPOCKET_segmented_pockets['up_aas'] = DEEPPOCKET_segmented_pockets['up_aas'].apply(sorted)
DEEPPOCKET_rescored_pockets['up_aas'] = DEEPPOCKET_rescored_pockets['up_aas'].apply(sorted)
P2RANK_CONS_pockets['up_aas'] = P2RANK_CONS_pockets['up_aas'].apply(sorted)
P2RANK_pockets['up_aas'] = P2RANK_pockets['up_aas'].apply(sorted)
FPOCKET_pockets['up_aas'] = FPOCKET_pockets['up_aas'].apply(sorted)
POCKETFINDER_pockets['up_aas'] = POCKETFINDER_pockets['up_aas'].apply(sorted)
LIGSITE_pockets['up_aas'] = LIGSITE_pockets['up_aas'].apply(sorted)
SURFNET_pockets['up_aas'] = SURFNET_pockets['up_aas'].apply(sorted)

ion_prop = read_from_pickle("./../DATA/LIGYSIS_human_sites_ion_prop_APRIL_2024.pkl")

max_dist_per_seg = read_from_pickle("./results/LIGYSIS_max_dist_per_seg.pkl")

prop_mono_per_site = read_from_pickle("./results/LIGYSIS_prop_mono_per_site.pkl")

LIGYSIS_sites.SEG = LIGYSIS_sites.SEG.astype(int)

LIGYSIS_sites["ion_prop"] = LIGYSIS_sites.set_index(['ACC', 'SEG', 'ID']).index.map(ion_prop)
LIGYSIS_sites["cluster_dist"] = LIGYSIS_sites.set_index('SEG_ID').index.map(max_dist_per_seg)
LIGYSIS_sites["monomer_prop"] = LIGYSIS_sites.set_index('SITE_NAME').index.map(prop_mono_per_site)

LIGYSIS_cols =                 ["origin", "rep_chain", "ID",                           "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas", "ion_prop", "cluster_dist", "monomer_prop"]
VNEGNN_cols =                  ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
IFSP_cols =                    ["origin", "rep_chain", "ID", "RANK",      "n_points",  "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
GRASP_cols =                   ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
PURESNET_cols =                ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"] # added score as n_aas. Nevertheless, RANK = ID.
DEEPPOCKET_segmented_cols =    ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
DEEPPOCKET_rescored_cols =     ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG",                "centre", "up_aas", "aas"] # don't have them as they are the same as fpocket ones.
P2RANK_CONS_cols =             ["origin", "rep_chain", "ID", "RANK",      "prob",      "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
P2RANK_cols =                  ["origin", "rep_chain", "ID", "RANK",      "prob",      "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
FPOCKET_cols =                 ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
POCKETFINDER_cols =            ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
LIGSITE_cols =                 ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]
SURFNET_cols =                 ["origin", "rep_chain", "ID", "RANK",      "score",     "n_aas", "RoG", "SASA", "VOL", "centre", "up_aas", "aas"]

len(LIGYSIS_sites.rep_chain.unique())

LIGYSIS_reps = sorted(list(LIGYSIS_sites.rep_chain.unique()))
#save_to_pickle(LIGYSIS_reps, "./results/LIGYSIS_3448_chains.pkl")

## Merging PUResNet data with PRANK-scored pockets

print(len(PURESNET_PRANK_pockets))

print(len(PURESNET_pockets))

PURESNET_PRANK_pockets_filt = PURESNET_PRANK_pockets.merge(
    PURESNET_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
PURESNET_PRANK_pockets_filt.n_aas = PURESNET_PRANK_pockets_filt.n_aas.astype(int)

PURESNET_PRANK_pockets_filt[PURESNET_cols].head(3)

## Merging PUResNet data with PRANK+Cons-scored pockets

PURESNET_PRANK_CONS_pockets_filt = PURESNET_PRANK_CONS_pockets.merge(
    PURESNET_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
PURESNET_PRANK_CONS_pockets_filt.n_aas = PURESNET_PRANK_CONS_pockets_filt.n_aas.astype(int)
#PURESNET_PRANK_CONS_pockets_filt.VOL = PURESNET_PRANK_CONS_pockets_filt.VOL.astype(int)

PURESNET_PRANK_CONS_pockets_filt[PURESNET_cols].head(3)

## Merging fpocket data with PRANK-rescored fpocket candidates

FPOCKET_PRANK_pockets_filt = FPOCKET_PRANK_pockets.merge(
    FPOCKET_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
FPOCKET_PRANK_pockets_filt.n_aas = FPOCKET_PRANK_pockets_filt.n_aas.astype(int)
#FPOCKET_PRANK_pockets_filt.VOL = FPOCKET_PRANK_pockets_filt.VOL.astype(int)

FPOCKET_PRANK_pockets_filt[FPOCKET_cols].head(3)

## Merging fpocket data with PRANK+Cons-rescored fpocket candidates

FPOCKET_PRANK_CONS_pockets_filt = FPOCKET_PRANK_CONS_pockets.merge(
    FPOCKET_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
FPOCKET_PRANK_CONS_pockets_filt.n_aas = FPOCKET_PRANK_CONS_pockets_filt.n_aas.astype(int)
#FPOCKET_PRANK_CONS_pockets_filt.VOL = FPOCKET_PRANK_CONS_pockets_filt.VOL.astype(int)

FPOCKET_PRANK_CONS_pockets_filt[FPOCKET_cols].head(3)

## Merging PocketFinder data with PRANK-scored pockets

POCKETFINDER_PRANK_pockets_filt = POCKETFINDER_PRANK_pockets.merge(
    POCKETFINDER_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
POCKETFINDER_PRANK_pockets_filt.n_aas = POCKETFINDER_PRANK_pockets_filt.n_aas.astype(int)
#POCKETFINDER_PRANK_pockets_filt.VOL = POCKETFINDER_PRANK_pockets_filt.VOL.astype(int)

POCKETFINDER_PRANK_pockets_filt[POCKETFINDER_cols].head(3)

## Merging PocketFinder data with PRANK+Cons-rescored pockets

POCKETFINDER_PRANK_CONS_pockets_filt = POCKETFINDER_PRANK_CONS_pockets.merge(
    POCKETFINDER_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
POCKETFINDER_PRANK_CONS_pockets_filt.n_aas = POCKETFINDER_PRANK_CONS_pockets_filt.n_aas.astype(int)
#POCKETFINDER_PRANK_CONS_pockets_filt.VOL = POCKETFINDER_PRANK_CONS_pockets_filt.VOL.astype(int)

POCKETFINDER_PRANK_CONS_pockets_filt[POCKETFINDER_cols].head(3)

## Merging Ligsite data with PRANK-scored pockets

LIGSITE_PRANK_pockets_filt = LIGSITE_PRANK_pockets.merge(
    LIGSITE_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
LIGSITE_PRANK_pockets_filt.n_aas = LIGSITE_PRANK_pockets_filt.n_aas.astype(int)
#LIGSITE_PRANK_pockets_filt.VOL = LIGSITE_PRANK_pockets_filt.VOL.astype(int)

LIGSITE_PRANK_pockets_filt[LIGSITE_cols].head(3)

## Merging Ligsite data with PRANK+Cons-scored pockets

LIGSITE_PRANK_CONS_pockets_filt = LIGSITE_PRANK_CONS_pockets.merge(
    LIGSITE_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
LIGSITE_PRANK_CONS_pockets_filt.n_aas = LIGSITE_PRANK_CONS_pockets_filt.n_aas.astype(int)
#LIGSITE_PRANK_CONS_pockets_filt.VOL = LIGSITE_PRANK_CONS_pockets_filt.VOL.astype(int)

LIGSITE_PRANK_CONS_pockets_filt[LIGSITE_cols].head(3)

## Merging Surfnet data with PRANK-rescored pockets

SURFNET_PRANK_pockets_filt = SURFNET_PRANK_pockets.merge(
    SURFNET_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
SURFNET_PRANK_pockets_filt.n_aas = SURFNET_PRANK_pockets_filt.n_aas.astype(int)
#SURFNET_PRANK_pockets_filt.VOL = SURFNET_PRANK_pockets_filt.VOL.astype(int)

SURFNET_PRANK_pockets_filt[SURFNET_cols].head(3)

## Merging Surfnet data with PRANK+Cons-scored pockets

SURFNET_PRANK_CONS_pockets_filt = SURFNET_PRANK_CONS_pockets.merge(
    SURFNET_pockets[['rep_chain', 'ID', 'n_aas', 'RoG', 'SASA', 'VOL', 'centre', 'up_aas', 'aas']],
    on=['rep_chain', 'ID'],
    how='left'
).query('n_aas == n_aas').reset_index(drop=True) #.dropna()
SURFNET_PRANK_CONS_pockets_filt.n_aas = SURFNET_PRANK_CONS_pockets_filt.n_aas.astype(int)
#SURFNET_PRANK_CONS_pockets_filt.VOL = SURFNET_PRANK_CONS_pockets_filt.VOL.astype(int)

SURFNET_PRANK_CONS_pockets_filt[SURFNET_cols].head(3)

### Filter out structures for which SIFTS mapping to UniProt residues in predicted sites fail

bad_reps = read_from_pickle("./results/OLD/other_errrors_reps.pkl")

no_sifts_vnegnn = read_from_pickle("./results/VNEGNN_no_sifts_TRANS.pkl")
no_sifts_ifsp = read_from_pickle("./results/IFSP_no_sifts_TRANS.pkl")
no_sifts_grasp = read_from_pickle("./results/GrASP_no_sifts_TRANS.pkl")
no_sifts_puresnet = read_from_pickle("./results/PURESNET_no_sifts.pkl")
no_sifts_deeppocket = read_from_pickle("./results/DeepPocket_no_sifts_TRANS.pkl")
no_sifts_p2rank_cons = read_from_pickle("./results/P2Rank_CONS_no_sifts_TRANS.pkl")
no_sifts_p2rank = read_from_pickle("./results/P2Rank_no_sifts_TRANS.pkl")
no_sifts_fpocket = read_from_pickle("./results/fpocket_no_sifts_TRANS.pkl")
no_sifts_pocketfinder = read_from_pickle("./results/POCKETFINDER_no_sifts.pkl")
no_sifts_ligsite = read_from_pickle("./results/LIGSITE_no_sifts.pkl")
no_sifts_surfnet = read_from_pickle("./results/SURFNET_no_sifts.pkl")

no_sifts_un = sorted(
    list(
        set(
            no_sifts_vnegnn + no_sifts_ifsp + no_sifts_grasp + no_sifts_puresnet + no_sifts_deeppocket + no_sifts_p2rank_cons + no_sifts_p2rank + no_sifts_fpocket + no_sifts_pocketfinder + no_sifts_ligsite + no_sifts_surfnet 
        )
    )
)
print(len(no_sifts_un))

LIGYSIS_sites_filt = LIGYSIS_sites.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
VNEGNN_pockets_filt = VNEGNN_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
IFSP_pockets_filt = IFSP_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
GRASP_pockets_filt = GRASP_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

PURESNET_PRANK_CONS_pockets_filt = PURESNET_PRANK_CONS_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
PURESNET_PRANK_pockets_filt = PURESNET_PRANK_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
PURESNET_pockets_filt = PURESNET_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

DEEPPOCKET_segmented_pockets_filt = DEEPPOCKET_segmented_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
DEEPPOCKET_rescored_pockets_filt = DEEPPOCKET_rescored_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

P2RANK_CONS_pockets_filt = P2RANK_CONS_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
P2RANK_pockets_filt = P2RANK_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

FPOCKET_PRANK_CONS_pockets_filt = FPOCKET_PRANK_CONS_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
FPOCKET_PRANK_pockets_filt = FPOCKET_PRANK_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
FPOCKET_pockets_filt = FPOCKET_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

POCKETFINDER_PRANK_CONS_pockets_filt = POCKETFINDER_PRANK_CONS_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
POCKETFINDER_PRANK_pockets_filt = POCKETFINDER_PRANK_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
POCKETFINDER_pockets_filt = POCKETFINDER_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

LIGSITE_PRANK_CONS_pockets_filt = LIGSITE_PRANK_CONS_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
LIGSITE_PRANK_pockets_filt = LIGSITE_PRANK_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
LIGSITE_pockets = LIGSITE_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

SURFNET_PRANK_CONS_pockets_filt = SURFNET_PRANK_CONS_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
SURFNET_PRANK_pockets_filt = SURFNET_PRANK_pockets_filt.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)
SURFNET_pockets_filt = SURFNET_pockets.query('rep_chain not in @no_sifts_un').copy().reset_index(drop = True)

save_to_pickle(no_sifts_un, "./results/ALL_no_sifts_TRANS.pkl")

LIGYSIS_sites_filt = LIGYSIS_sites_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
VNEGNN_pockets_filt = VNEGNN_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
IFSP_pockets_filt = IFSP_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
GRASP_pockets_filt = GRASP_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

PURESNET_PRANK_CONS_pockets_filt = PURESNET_PRANK_CONS_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
PURESNET_PRANK_pockets_filt = PURESNET_PRANK_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
PURESNET_pockets_filt = PURESNET_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

DEEPPOCKET_segmented_pockets_filt = DEEPPOCKET_segmented_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
DEEPPOCKET_rescored_pockets_filt = DEEPPOCKET_rescored_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

P2RANK_CONS_pockets_filt = P2RANK_CONS_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
P2RANK_pockets_filt = P2RANK_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

FPOCKET_PRANK_CONS_pockets_filt = FPOCKET_PRANK_CONS_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
FPOCKET_PRANK_pockets_filt = FPOCKET_PRANK_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
FPOCKET_pockets_filt = FPOCKET_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

POCKETFINDER_PRANK_CONS_pockets_filt = POCKETFINDER_PRANK_CONS_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
POCKETFINDER_PRANK_pockets_filt = POCKETFINDER_PRANK_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
POCKETFINDER_pockets_filt = POCKETFINDER_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

LIGSITE_PRANK_CONS_pockets_filt = LIGSITE_PRANK_CONS_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
LIGSITE_PRANK_pockets_filt = LIGSITE_PRANK_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
LIGSITE_pockets = LIGSITE_pockets.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

SURFNET_PRANK_CONS_pockets_filt = SURFNET_PRANK_CONS_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
SURFNET_PRANK_pockets_filt = SURFNET_PRANK_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)
SURFNET_pockets_filt = SURFNET_pockets_filt.query('rep_chain not in @bad_reps').copy().reset_index(drop = True)

len(LIGYSIS_sites_filt.ACC.unique())

LIGYSIS_site_reps = LIGYSIS_sites_filt.rep_chain.unique().tolist()
print(len(LIGYSIS_site_reps)) #2775

VNEGNN_pockets_filt = VNEGNN_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
IFSP_pockets_filt = IFSP_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
GRASP_pockets_filt = GRASP_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

PURESNET_PRANK_CONS_pockets_filt = PURESNET_PRANK_CONS_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
PURESNET_PRANK_pockets_filt = PURESNET_PRANK_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
PURESNET_pockets_filt = PURESNET_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

DEEPPOCKET_segmented_pockets_filt = DEEPPOCKET_segmented_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)                                                                                                                            
DEEPPOCKET_rescored_pockets_filt = DEEPPOCKET_rescored_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

P2RANK_CONS_pockets_filt = P2RANK_CONS_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
P2RANK_pockets_filt = P2RANK_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

FPOCKET_PRANK_CONS_pockets_filt = FPOCKET_PRANK_CONS_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
FPOCKET_PRANK_pockets_filt = FPOCKET_PRANK_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
FPOCKET_pockets_filt = FPOCKET_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

POCKETFINDER_PRANK_CONS_pockets_filt = POCKETFINDER_PRANK_CONS_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
POCKETFINDER_PRANK_pockets_filt = POCKETFINDER_PRANK_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
POCKETFINDER_pockets_filt = POCKETFINDER_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

LIGSITE_PRANK_CONS_pockets_filt = LIGSITE_PRANK_CONS_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
LIGSITE_PRANK_pockets_filt = LIGSITE_PRANK_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
LIGSITE_pockets = LIGSITE_pockets.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

SURFNET_PRANK_CONS_pockets_filt = SURFNET_PRANK_CONS_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
SURFNET_PRANK_pockets_filt = SURFNET_PRANK_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)
SURFNET_pockets_filt = SURFNET_pockets_filt.query('rep_chain in @LIGYSIS_site_reps').copy().reset_index(drop = True)

combined_df = pd.concat(
    [
        LIGYSIS_sites_filt[LIGYSIS_cols],
        VNEGNN_pockets_filt[VNEGNN_cols],
        IFSP_pockets_filt[IFSP_cols],
        GRASP_pockets_filt[GRASP_cols],
        
        PURESNET_PRANK_CONS_pockets_filt[PURESNET_cols], 
        PURESNET_PRANK_pockets_filt[PURESNET_cols],
        PURESNET_pockets_filt[PURESNET_cols],
        
        DEEPPOCKET_segmented_pockets_filt[DEEPPOCKET_segmented_cols],
        DEEPPOCKET_rescored_pockets_filt[DEEPPOCKET_rescored_cols],
        
        P2RANK_CONS_pockets_filt[P2RANK_CONS_cols],
        P2RANK_pockets_filt[P2RANK_cols],
        
        FPOCKET_PRANK_CONS_pockets_filt[FPOCKET_cols], 
        FPOCKET_PRANK_pockets_filt[FPOCKET_cols], 
        FPOCKET_pockets_filt[FPOCKET_cols],
        
        POCKETFINDER_PRANK_CONS_pockets_filt[POCKETFINDER_cols], 
        POCKETFINDER_PRANK_pockets_filt[POCKETFINDER_cols], 
        POCKETFINDER_pockets_filt[POCKETFINDER_cols],  
        
        LIGSITE_PRANK_CONS_pockets_filt[LIGSITE_cols], 
        LIGSITE_PRANK_pockets_filt[LIGSITE_cols], 
        LIGSITE_pockets[LIGSITE_cols],  
        
        SURFNET_PRANK_CONS_pockets_filt[SURFNET_cols], 
        SURFNET_PRANK_pockets_filt[SURFNET_cols], 
        SURFNET_pockets_filt[SURFNET_cols],  
    ]
).reset_index(drop = True)

combined_df_filt = combined_df.query('n_aas > 0').copy().reset_index(drop = True)

len(combined_df_filt.rep_chain.unique().tolist())

combined_df_filt.query('VOL != VOL').origin.value_counts()

combined_df_filt.to_pickle("./results/ALL_PREDS_COMBINED_RIGHT_EXTENDED_DEF.pkl")
